+++ App.js
@@ -1,14 +1,16 @@
 import { NavigationContainer } from '@react-navigation/native';
 import React from 'react';
+import AppContextProvider from './App/Context/AppContext';
 import DrawerNavigation from './App/Routes/DrawerNavigation';
 
 export default function App() {
 
   return (
-    
-    <NavigationContainer>
-      <DrawerNavigation />
-    </NavigationContainer>
+    <AppContextProvider>
+      <NavigationContainer>
+        <DrawerNavigation />
+      </NavigationContainer>
+    </AppContextProvider>
   );
 }