+++ App/Routes/DrawerNavigation.js
@@ -1,7 +1,6 @@
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import React from 'react';
-import HomeScreen from '../Screens/HomeScreen';
-import Income from '../Screens/Income';
+import HomeStackNavigation from './HomeStackNavigation';
 
 
 const Drawer = createDrawerNavigator()
@@ -9,8 +8,7 @@ const Drawer = createDrawerNavigator()
 function DrawerNavigation() {
     return (
         <Drawer.Navigator initialRouteName="Home">
-            <Drawer.Screen name="Home" component={HomeScreen} />
-            <Drawer.Screen name="Income" component={Income} />
+            <Drawer.Screen name="Home" component={HomeStackNavigation} />
         </Drawer.Navigator>
     )
 }