+++ App/Screens/HomeScreen.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { StatusBar, StyleSheet, Text, View } from 'react-native';
+import { Image, StatusBar, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import Colors from '../Config/Colors';
 
 
-export default function HomeScreen() {
+export default function HomeScreen({ navigation }) {
     return (
         <View style={styles.container}>
             <StatusBar
@@ -23,6 +23,66 @@ export default function HomeScreen() {
                 <Text style={styles.cashTitle}>
                     Cash Available
                 </Text>
+            </View>
+            <View style={styles.catagoryWrapper}>
+                <Text style={styles.catagoryHeading}>
+                    Your main catagories
+                </Text>
+                <View style={styles.catagoryIconContainer}>
+                    <TouchableOpacity style={styles.catagory} onPress={()=>navigation.push('Income')}>
+                        <View style={styles.catagoryIcon}>
+                            <Image
+                                style={{ height: 45, width: 45 }}
+                                source={require('../assets/icons/catagory/income.png')} />
+                        </View>
+                        <View style={styles.catagoryName}>
+                            <Text
+                                style={{ fontSize: 16, textAlign: "center", fontWeight: "600" }}>
+                                Income
+                            </Text>
+                        </View>
+                    </TouchableOpacity>
+                    <TouchableOpacity style={styles.catagory} onPress={()=>navigation.push('Deposit')}>
+                        <View style={styles.catagoryIcon}>
+                            <Image
+                                style={{ height: 45, width: 45 }}
+                                source={require('../assets/icons/catagory/deposit.png')} />
+                        </View>
+                        <View style={styles.catagoryName}>
+                            <Text
+                                style={{ fontSize: 16, textAlign: "center", fontWeight: "600" }}>
+                                Deposit
+                            </Text>
+                        </View>
+                    </TouchableOpacity>
+                    <TouchableOpacity style={styles.catagory} onPress={()=>navigation.push('Expense')}>
+                        <View style={styles.catagoryIcon}>
+                            <Image
+                                style={{ height: 45, width: 45 }}
+                                source={require('../assets/icons/catagory/expense.png')} />
+                        </View>
+                        <View style={styles.catagoryName}>
+                            <Text
+                                style={{ fontSize: 16, textAlign: "center", fontWeight: "600" }}>
+                                Expense
+                            </Text>
+                        </View>
+                    </TouchableOpacity>
+                    <TouchableOpacity style={styles.catagory} onPress={()=>navigation.push('Loan')}>
+                        <View style={styles.catagoryIcon}>
+                            <Image
+                                style={{ height: 45, width: 45 }}
+                                source={require('../assets/icons/catagory/loan.png')} />
+                        </View>
+                        <View style={styles.catagoryName}>
+                            <Text
+                                style={{ fontSize: 16, textAlign: "center", fontWeight: "600" }}>
+                                Loan
+                            </Text>
+                        </View>
+                    </TouchableOpacity>
+                </View>
+
             </View>
         </View>
     )
@@ -45,7 +105,9 @@ const styles = StyleSheet.create({
         paddingTop: 20,
         paddingBottom: 20,
         justifyContent: "center",
+        backgroundColor: "#6B7AA1",
         borderRadius: 10,
+
     },
     cashTitle: {
         textAlign: "center",
@@ -57,5 +119,36 @@ const styles = StyleSheet.create({
         color: Colors.cashAmount,
         fontSize: 35,
 
+    },
+    catagoryWrapper: {
+        alignItems: "center",
+        marginTop: 10,
+
+    },
+    catagoryHeading: {
+        fontSize: 20,
+        color: Colors.catagoryHeading
+    },
+    catagoryIconContainer: {
+        marginTop: 10,
+        flexDirection: "row",
+        justifyContent: "space-around",
+        width: "90%",
+    },
+    catagory: {
+        backgroundColor: "#CEE5D0",
+        padding: 5,
+        alignItems: "center",
+        borderRadius: 15,
+
+    },
+    catagoryIcon: {
+        backgroundColor: "#fff",
+        padding: 5,
+        borderRadius: 10,
+    },
+    catagoryName: {
+        // padding: 3,
+        borderRadius: 10,
     },
 })